chore(config): split Token Rhythm preset files
Problem: the provider preset implementation and test file exceeded repolint's incremental size budgets.\n\nFix: move Token Rhythm model declarations, overrides, and integration coverage into dedicated files without changing behavior.\n\nVerification: go run ./tools/repolint; go test ./internal/config ./internal/provider; git diff --check
